memory manager: refactor
Separate definitions from function code by creating header file for statics
managment.
